% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGWAStoManifest.R
\name{addGWAStoManifest}
\alias{addGWAStoManifest}
\title{addGWAStoManifest}
\usage{
addGWAStoManifest(
  short = c("UniqueTraitName"),
  n = c(10000),
  filename = c("?"),
  year = c("?"),
  trait = c("?"),
  DOI = c("?"),
  genomeBuild = c("?"),
  traitType = c("?"),
  rawSNPs = c("?"),
  finalModelSNPs = c("?"),
  modelRunningTime = c("?"),
  usedRefSet = c("?"),
  processed = c(0),
  FORCE = FALSE
)
}
\value{
This function will add a GWAS with all its needed information to the Manifest file
}
\description{
addGWAStoManifest
}
\examples{
addGWAStoManifest(short=c("Height"),	n=c(10000), filename=c("C:/Path/To/GWAS_file"), year=c("2020"), trait=c("Height"), DOI=c("?"),	genomeBuild = c("hg19"), traitType = c("CON"), rawSNPs = c("?"), finalModelSNPs = c("?"), modelRunningTime = c("?"), usedRefSet = c("?"), processed=c(0))
The most important attributes are "short", "n", "filename" and "traitType". "traitType" should be "CAT" or "CON".
}
