% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPGS_LDAK.R
\name{calcPGS_LDAK}
\alias{calcPGS_LDAK}
\title{calcPGS_LDAK}
\usage{
calcPGS_LDAK(Trait = NA, Model = "bayesr")
}
\value{
This function will perform the generation of the PGS model, of the given GWAS.
}
\description{
calcPGS_LDAK
}
\examples{
calcPGS_LDAK(Trait = "Height",Model = "bayesr")
}
