% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_all_PRS.R
\name{collect_all_PRS}
\alias{collect_all_PRS}
\title{collect_all_PRS}
\usage{
collect_all_PRS(cohort = NA, Model = "bayesr", Trait = NA)
}
\arguments{
\item{cohort}{Enter the file name (not the path) of the cohort the polygenic scores need collecing for.}

\item{Model}{Enter the model to collect the polygenic scores for specified trait.}

\item{Trait}{Enter the unique GWAS trait ID of interest to collect the polygenic scores using specified model.}
}
\value{
This function will return all the generated PGSs that were ran on a specific cohort, it will automatically load all and prompt a dataframe with PGS as columns and samples as rows.
}
\description{
collect_all_PRS
}
\examples{
all_PGS = collect_all_PRS(cohort = "examplecohort")
}
