% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportManifest.R
\name{exportManifest}
\alias{exportManifest}
\title{exportManifest}
\usage{
exportManifest(Path = NA)
}
\arguments{
\item{Path}{The folder path to save the manifest in. Automatically generates a tab-delimited file called 'Manifest.tsv' in specified folder.}
}
\value{
This function will export the manifest used in this package to the specified folder. The manifest is the heart of the package, this must never get corrupted.
}
\description{
exportManifest
}
\examples{
exportManifest(Path = "C:/Users/Default/Downloads/")
}
