% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTraits.R
\name{getTraits}
\alias{getTraits}
\title{getManifest}
\usage{
getTraits()
}
\value{
This fucntion is for vizualization purposes, it shows all the 'shorts' and their descriptions.
}
\description{
getManifest
}
\examples{
getTraits()
}
