% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importManifest.R
\name{importManifest}
\alias{importManifest}
\title{importManifest}
\usage{
importManifest(Path = NA)
}
\arguments{
\item{Path}{The specific file path to get the manifest from. Must be a tab-delimited file. See formatting example at exportManifest().}
}
\value{
This function will import the manifest used in this package from the specified file. The manifest is the heart of the package, this must never get corrupted.
}
\description{
importManifest
}
\examples{
importManifest(Path = "C:/Users/Default/Downloads/Manifest.tsv")
}
