% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installDependenciesAndData.R
\name{installDependenciesAndData}
\alias{installDependenciesAndData}
\title{installDependenciesAndData}
\usage{
installDependenciesAndData()
}
\value{
This function is essential in the workings for the PRSMultiTrait package, as it installs, checks, and sets everything up.
}
\description{
installDependenciesAndData
}
\examples{
installDependenciesAndData()
}
