% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyGWASinManifest.R
\name{modifyGWASinManifest}
\alias{modifyGWASinManifest}
\title{modifyGWASinManifest}
\usage{
modifyGWASinManifest(
  short = c("UniqueTraitName"),
  n = NA,
  filename = NA,
  year = NA,
  trait = NA,
  DOI = NA,
  genomeBuild = NA,
  traitType = NA,
  rawSNPs = NA,
  finalModelSNPs = NA,
  modelRunningTime = NA,
  usedRefSet = NA,
  processed = NA,
  FORCE = FALSE
)
}
\description{
modifyGWASinManifest
}
