% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPGSApp.R
\name{runPGSApp}
\alias{runPGSApp}
\title{runPGSApp}
\usage{
runPGSApp()
}
\value{
This function will start the shiny interface to use this package. This way interacting with the package will be faster, more effecient, and less error-prone.
}
\description{
runPGSApp
}
\examples{
runPGSApp()
}
