% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveManifest.R
\name{saveManifest}
\alias{saveManifest}
\title{saveManifest}
\usage{
saveManifest()
}
\value{
This function will save the manifest to its dedicated file / enviroment.
}
\description{
saveManifest
}
\examples{
saveManifest()
}
