% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGWAStoManifest.R
\name{addGWAStoManifest}
\alias{addGWAStoManifest}
\title{addGWAStoManifest}
\usage{
addGWAStoManifest(
  short = c("UniqueTraitName"),
  n = c(10000),
  filename = c("?"),
  year = c("?"),
  trait = c("?"),
  DOI = c("?"),
  genomeBuild = c("?"),
  traitType = c("?"),
  rawSNPs = c("?"),
  finalModelSNPs = c("?"),
  modelRunningTime = c("?"),
  usedRefSet = c("?"),
  processed = c(0),
  FORCE = FALSE
)
}
\description{
addGWAStoManifest
}
