% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPGS_LDAK.R
\name{calcPGS_LDAK}
\alias{calcPGS_LDAK}
\title{calcPGS_LDAK}
\usage{
calcPGS_LDAK(Trait = NA, Model = "bayesr")
}
\arguments{
\item{Trait}{Enter the unique GWAS trait ID of interest to be used in modelling the PGM. Can be observed by getTraits().}

\item{Model}{Enter the model to be used in modelling the PGM.}
}
\value{
This function will perform the generation of the PGS model, of the given GWAS.
}
\description{
calcPGS_LDAK
}
\examples{
calcPGS_LDAK(Trait = "Height",Model = "bayesr")
}
