% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importManifest.R
\name{importManifest}
\alias{importManifest}
\title{importManifest}
\usage{
importManifest(Path = NA)
}
\description{
importManifest
}
\examples{
importManifest(Path = "C:/Users/p70072451/Downloads/Manifest.tsv")
}
