% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyGWASinManifest.R
\name{modifyGWASinManifest}
\alias{modifyGWASinManifest}
\title{modifyGWASinManifest}
\usage{
modifyGWASinManifest(
  short = c("UniqueTraitName"),
  n = NA,
  filename = NA,
  year = NA,
  trait = NA,
  DOI = NA,
  genomeBuild = NA,
  traitType = NA,
  rawSNPs = NA,
  finalModelSNPs = NA,
  modelRunningTime = NA,
  usedRefSet = NA,
  processed = NA,
  FORCE = FALSE
)
}
\value{
This function will alter an existing GWAS trait in the manifest. This is based on using the 'short' as a method to select which GWAS is needed to be changed, and any other non-NA argument is changed into its alteration.
}
\description{
modifyGWASinManifest
}
\examples{
modifyGWASinManifest(short=c("Height"), n = 12345)
}
