% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predPRS.R
\name{predPRS}
\alias{predPRS}
\title{predPRS}
\usage{
predPRS(
  bfile = NA,
  Trait = NA,
  Model = "bayesr",
  OverlapSNPsOnly = FALSE,
  Force = FALSE
)
}
\value{
This fucntion will run the calculation of the PGS of the given GWAS, it will automatically save these PGS which can be extracted with collect_all_PRS(). Make sure the bfiles are using a chr:bp ids and are using wsl-pathing notation in the fucntion.
}
\description{
predPRS
}
\examples{
predPRS(bfile = wslpath("C:/path/to/examplecohort"), Trait = "Height")
}
