% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeGWASfromManifest.R
\name{removeGWASfromManifest}
\alias{removeGWASfromManifest}
\title{removeGWASfromManifest}
\usage{
removeGWASfromManifest(trait = c("UniqueTraitName"), FORCE = FALSE)
}
\description{
removeGWASfromManifest
}
